Rename file news-item-card.vue to NewsItemCard.vue
